move clustering stuff into clustering folders
