+++ tests/clustering/math/test_compute_rank.py
