+++ tests/clustering/math/test_perm_invariant_hamming.py
